Reactor & update dependencies
