+++ client/Cargo.toml
@@ -18,7 +18,7 @@ path = "src/lib.rs"
 [dependencies]
 spaces_wallet = { path = "../wallet" }
 spaces_protocol = { path = "../protocol", version = "*", features = ["std"]}
-spacedb = { git = "https://github.com/spacesprotocol/spacedb", version = "0.0.5" }
+spacedb = { git = "https://github.com/spacesprotocol/spacedb", version = "0.0.6" }
 
 tokio = { version = "1.37.0", features = ["signal"] }
 ctrlc = "3.4.4"
@@ -26,13 +26,12 @@ anyhow = "1.0.86"
 clap = { version = "4.5.6", features = ["derive", "env"] }
 log = "0.4.21"
 serde = { version = "1.0.200", features = ["derive"] }
-toml = "0.8.14"
 hex = "0.4.3"
 jsonrpsee = { version = "0.22.5", features = ["server", "http-client", "macros"] }
 directories = "5.0.1"
 env_logger = "0.11.3"
 serde_json = "1.0.116"
-bincode = {version = "2.0.0-rc.3", features = ["serde", "derive"]}
+bincode = {version = "2.0.1", features = ["serde", "derive"]}
 base64 = "0.22.1"
 futures = "0.3.30"
 reqwest = { version = "0.12.5", default-features = false, features = ["json", "blocking", "rustls-tls"] }
