+++ client/src/config.rs
@@ -1,7 +1,4 @@
 use std::{
-    collections::{HashMap, HashSet},
-    env,
-    ffi::OsString,
     fmt::Display,
     fs,
     net::{IpAddr, SocketAddr},
@@ -9,7 +6,6 @@ use std::{
 };
 
 use clap::{
-    error::{ContextKind, ContextValue},
     ArgGroup, Parser, ValueEnum,
 };
 use directories::ProjectDirs;
@@ -17,7 +13,6 @@ use jsonrpsee::core::Serialize;
 use log::error;
 use serde::Deserialize;
 use spaces_protocol::bitcoin::Network;
-use toml::Value;
 
 use crate::{
     source::{BitcoinRpc, BitcoinRpcAuth},
@@ -37,9 +32,6 @@ const RPC_OPTIONS: &str = "RPC Server Options";
 ))]
 #[command(args_override_self = true, author, version, about, long_about = None)]
 pub struct Args {
-    /// Path to a configuration file
-    #[arg(long, env = "SPACED_CONFIG")]
-    config: Option<PathBuf>,
     #[arg(long, env = "SPACED_BLOCK_INDEX", default_value = "false")]
     block_index: bool,
     #[arg(long, env = "SPACED_DATA_DIR")]
@@ -109,23 +101,9 @@ impl ExtendedNetwork {
 impl Args {
     /// Configures spaced node by processing command line arguments
     /// and configuration files
-    pub async fn configure(shutdown: tokio::sync::broadcast::Receiver<()>) -> anyhow::Result<Spaced> {
-        let mut args = Args::merge_args_config(None);
+    pub async fn configure(args: Vec<String>, shutdown: tokio::sync::broadcast::Receiver<()>) -> anyhow::Result<Spaced> {
+        let mut args =  Args::try_parse_from(args)?;
         let default_dirs = get_default_node_dirs();
-
-        // Update from a configuration file if it exists
-        args = match args.config {
-            None => {
-                let default_path = default_dirs.config_dir().join("Spaces.toml");
-                args.config = Some(default_path.clone());
-
-                match default_path.exists() {
-                    true => Args::merge_args_config(Some(default_path)),
-                    false => args,
-                }
-            }
-            Some(user_specified_path) => Args::merge_args_config(Some(user_specified_path)),
-        };
 
         if args.bitcoin_rpc_url.is_none() {
             args.bitcoin_rpc_url = Some(default_bitcoin_rpc_url(&args.chain).to_string())
@@ -227,40 +205,6 @@ impl Args {
             cbf: args.bitcoin_rpc_light
         })
     }
-
-    /// Merges configuration file if set and command line arguments (latter takes priority)
-    fn merge_args_config(config_file: Option<PathBuf>) -> Self {
-        let mut config_args_keys = None;
-        let config_args = match load_args(config_file) {
-            None => Args::try_parse_from(env::args_os()),
-            Some((config_args, keys)) => {
-                config_args_keys = Some(keys);
-                let cmd_args = env::args_os().collect::<Vec<OsString>>();
-                let config_args: Vec<OsString> = config_args.iter().map(|s| s.into()).collect();
-                let all_args = cmd_args
-                    .iter()
-                    .take(1)
-                    .chain(config_args.iter())
-                    .chain(cmd_args.iter().skip(1));
-                Args::try_parse_from(all_args)
-            }
-        };
-
-        let args = match config_args {
-            Ok(args) => args,
-            Err(mut err) => {
-                if let Some(config_args_keys) = config_args_keys {
-                    err = style_config_error(err, config_args_keys);
-                    error!("{}", err.to_string().replace("argument", "option"));
-                    safe_exit(err.exit_code());
-                } else {
-                    err.exit()
-                }
-            }
-        };
-
-        args
-    }
 }
 
 fn get_default_node_dirs() -> ProjectDirs {
@@ -288,92 +232,6 @@ pub fn default_bitcoin_rpc_url(network: &ExtendedNetwork) -> &'static str {
         ExtendedNetwork::Testnet => "http://127.0.0.1:18332",
         ExtendedNetwork::Regtest => "http://127.0.0.1:18443",
     }
-}
-
-fn toml_value_to_string(value: Value) -> Option<String> {
-    match value {
-        Value::String(v) => Some(v),
-        Value::Integer(i) => Some(i.to_string()),
-        Value::Float(f) => Some(f.to_string()),
-        Value::Boolean(b) => Some(b.to_string()),
-        Value::Datetime(d) => Some(d.to_string()),
-        Value::Array(_) => None,
-        Value::Table(_) => None,
-    }
-}
-
-fn load_args(path: Option<PathBuf>) -> Option<(Vec<String>, HashSet<String>)> {
-    let path = match path {
-        None => return None,
-        Some(p) => p,
-    };
-
-    let mut config_args = HashSet::new();
-    if let Ok(config_str) = fs::read_to_string(path.clone()) {
-        let config: HashMap<String, toml::Value> = toml::from_str(&config_str).expect("parse");
-        let mut args = Vec::new();
-        for (key, value) in config {
-            let value = match toml_value_to_string(value) {
-                None => continue,
-                Some(v) => v,
-            };
-            let arg = format!("--{}", key.replace('_', "-"));
-            config_args.insert(arg.clone());
-            args.push(arg);
-            args.push(value);
-        }
-        return Some((args, config_args));
-    }
-
-    error!("could not read configuration at {}", path.to_str().unwrap());
-    safe_exit(1);
-}
-
-fn maybe_style_arg(str: &str, set: &HashSet<String>) -> String {
-    let arg_name = str.split_once(' ');
-    if let Some((arg_name, _)) = arg_name {
-        if set.contains(arg_name) {
-            return str
-                .replace("--", "config.")
-                .replace('-', "_")
-                .replace('<', "= <")
-                .replace('>', ">");
-        }
-    }
-
-    return str.to_string();
-}
-
-fn style_config_error(err: clap::Error, config_args: HashSet<String>) -> clap::Error {
-    let mut new_error = clap::Error::new(err.kind());
-    for (ck, cv) in err.context() {
-        match ck {
-            ContextKind::Usage
-            | ContextKind::Suggested
-            | ContextKind::SuggestedCommand
-            | ContextKind::SuggestedSubcommand
-            | ContextKind::Custom => {
-                // skip
-            }
-            _ => {
-                let new_cv = match cv {
-                    ContextValue::String(str) => {
-                        ContextValue::String(maybe_style_arg(str, &config_args))
-                    }
-                    ContextValue::Strings(strings) => {
-                        let mut strs = strings.clone();
-                        strs[0] = maybe_style_arg(&strs[0], &config_args);
-                        ContextValue::Strings(strs)
-                    }
-                    _ => cv.clone(),
-                };
-
-                new_error.insert(ck, new_cv);
-            }
-        }
-    }
-
-    new_error
 }
 
 impl Display for ExtendedNetwork {
