+++ client/src/lib.rs
@@ -19,6 +19,7 @@ pub mod store;
 pub mod spaces;
 pub mod wallets;
 mod cbf;
+pub mod app;
 
 fn std_wait<F>(mut predicate: F, wait: Duration)
 where
