+++ protocol/Cargo.toml
@@ -8,7 +8,7 @@ bitcoin = { version = "0.32.2", features = ["base64", "serde"], default-features
 log = "0.4.14"
 
 ## optional features
-bincode = { version = "2.0.0-rc.3", features = [ "derive", "serde", "alloc" ], default-features = false, optional = true }
+bincode = { version = "2.0.1", features = [ "derive", "serde", "alloc" ], default-features = false, optional = true }
 serde = { version = "^1.0", features = ["derive"], default-features = false, optional = true }
 
 [dev-dependencies]
