+++ wallet/Cargo.toml
@@ -13,7 +13,7 @@ anyhow = "1.0.80"
 bech32 = "0.11.0"
 serde = { version = "^1.0", features = ["derive"] }
 serde_json = "1.0"
-bincode = { version = "2.0.0-rc.3", features = ["serde"] }
+bincode = { version = "2.0.1", features = ["serde"] }
 jsonrpc = "0.18.0"
 ctrlc = "3.4.4"
 hex = "0.4.3"
